Added default root and /indi path, documented webclient, added simple gui example, cleaned up config, removed templates not needed, reoragnized js files and separated codebases for readability.
