--- example_clients/pyindi-panel.html
@@ -1,70 +1,0 @@
-<!doctype html>
-<html lang="en">
-	<head>
-		<!-- Required meta tags -->
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<link rel="apple-touch-icon" sizes="180x180" href="/indi/static/favicon/apple-touch-icon.png">
-		<link rel="shortcut icon" type="image/png" sizes="32x32" href="">
-		<link rel="shortcut icon" type="image/png" sizes="16x16" href="/indi/static/favicon/favicon-16x16.png">
-		<link rel="manifest" href="/indi/static/favicon/site.webmanifest">
-		<link href="/indi/static/fontawesome/css/all.css" rel="stylesheet">
-		<title>pyINDI Panel</title>
-		<!-- Load pyINDI scripts -->
-		<script src="/indi/static/js/indi.js"></script>
-		<script src="/indi/static/js/maps-indi.js"></script>
-		<script src="/indi/static/js/constants.js"></script>
-		<script src="/indi/static/js/config.js"></script>
-		<!-- Load pyINDI styling -->
-		<link rel="stylesheet" href="/indi/static/css/indi.css">
-		<script>
-			/* Constants */
-			var CUSTOM_GUI = false; // Triggers to true if html is in body tag below
-
-
-			/**
-			* Runs on load. Builds the default properties using a callback.
-			*/
-			const main = () => {
-				// Tornado way of getting list
-				var devices = {% raw json_encode(device_name) %};
-				builder.init(devices);
-
-				return;
-			}
-
-
-			/**
-			* This function will be called whenever an INDI property with the device 
-			*	{{ device_name }} is received. This is where users should modify the
-			*	code to append to specific objects.
-			*
-			*	If wanting to use a custom GUI, enable in the configuration and provide the correct syntax for selecting 
-			* elements.
-			* @param {Object} indi Contains all information about INDI property
-			*/
-			const handleProperty = (indi) => {
-				if (indi.op === IndiOp.DEFINITION) {
-					if (!UserConfig.CUSTOM_GUI) {
-						updater.handle(indi);
-					}
-					else {
-						var vector = updater.custom(indi);
-						// If returned vector is null, bad selector
-						if (!vector) {
-							return;
-						}
-						// Handle vector for the first time here
-					}
-				}
-				// Handle vector every time it comes in
-				var vector = updater.vector(indi);
-
-        return;
-			}
-		</script>
-	</head>
-	<body>
-	</body>
-	<script>document.body.addEventListener("load", main());</script>
-</html>
