+++ pyindi/client.py
@@ -1,4 +1,4 @@
-#1/usr/bin/env python3
+#!/usr/bin/env python3
 
 # Python imports
 import asyncio
