--- pyindi/www/static/js/config.js
@@ -1,12 +1,0 @@
-/**
- * Configuration object that users can modify 
- * @enum
- */
-const UserConfig = {
-  CUSTOM_GUI: false, //If true, using custom GUI layout and will not build default
-  BLINKING_BUSY_ALERT_LIGHTS: false, // If true, INDI states "busy" and "alert" blink
-  LOGGING_WELCOME_MESSAGE: "INDI messages & logs displayed below", // Message to display at start of log
-  INDI_CONSOLE_DEBUG: true, // Prints INDI mesages to the console
-}
-
-
