+++ pyindi/www/templates/index.html
@@ -1,0 +1,39 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		{% raw pyindi_head %}
+		<title>{{ title }}</title>
+		<script>
+			/**
+			* Runs on load. Builds the default properties using a callback.
+			*/
+			const main = () => {
+				var devices = {% raw json_encode(devices) %};
+				initialize(devices, customGui=false);
+				return;
+			}
+
+			/**
+			* This function will be called whenever an INDI property with the device 
+			*	{{ devices }} is received. This is where users should modify the
+			*	code to append to specific objects.
+			*
+			*	If wanting to use a custom GUI, enable in the configuration and provide the correct syntax for selecting 
+			* elements.
+			* @param {Object} indi Contains all information about INDI property
+			*/
+			const handleProperty = (indi) => {
+				if (indi.op === IndiOp.DEFINITION) {
+					updater.handle(indi);
+					// Handle vector for the first time here
+				}
+				// Handle vector every time it comes in
+				var vector = updater.vector(indi);
+        return;
+			}
+		</script>
+	</head>
+	<body>
+	</body>
+	<script>document.body.addEventListener("load", main());</script>
+</html>
